requirements.txt
ultralytics
opencv-python
numpy
pyyaml
tqdm

